+++ irodsinfo/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class IrodsinfoConfig(AppConfig):
+    name = 'irodsinfo'